refactor(cfd-1d): Decompose hemolysis/mod.rs into models and dynamics leaf modules
Part of ATLAS-ARCH-007.
