+++ crates/cfd-1d/src/physics/hemolysis/mod.rs
@@ -35,8 +35,8 @@
 //! ```
 
 pub mod acoustic_radiation;
-
-use aequitas::systems::si::quantities::{Pressure, Time};
+mod dynamics;
+mod models;
 
 // ── Giersiepen model constants (re-exported from cfd-core SSOT) ───────────────
 //
@@ -65,357 +65,21 @@ pub use cfd_core::physics::hemolysis::GIERSIEPEN_MILLIFLUIDIC_STRESS as GIERSIEP
 pub use cfd_core::physics::hemolysis::CAVITATION_HI_SLOPE;
 
 // ── Public API ────────────────────────────────────────────────────────────────
-
-/// Giersiepen (1990) haemolysis index: `HI = C · t^α · τ^β`.
-///
-/// Thin wrapper delegating to [`cfd_core::physics::hemolysis::HemolysisModel::giersiepen_millifluidic`]
-/// so that the single source of truth for all model constants lives in cfd-core.
-/// Returns `0.0` for non-positive inputs (no damage when shear or time is
-/// zero or negative) and `0.0` on conversion error (never expected in practice).
-///
-/// # Arguments
-///
-/// * `shear`    — wall shear stress as an Aequitas `Pressure`
-/// * `duration` — exposure duration as an Aequitas `Time`
-///
-/// # Example
-///
-/// ```
-/// use aequitas::systems::si::quantities::{Pressure, Time};
-/// use cfd_1d::giersiepen_hi;
-/// let hi = giersiepen_hi(Pressure::from_base(50.0), Time::from_base(0.01));
-/// assert!((hi - 2.578_444_864_181_722_3e-3).abs() < 1e-15);
-/// ```
-#[inline]
-#[must_use]
-pub fn giersiepen_hi(shear: Pressure, duration: Time) -> f64 {
-    cfd_core::physics::hemolysis::HemolysisModel::giersiepen_millifluidic()
-        .damage_index(shear.into_base(), duration.into_base())
-        .unwrap_or(0.0)
-}
-
-/// Amplify a baseline Giersiepen HI by the local cavitation potential.
-///
-/// Thin wrapper delegating to [`cfd_core::physics::hemolysis::HemolysisModel::cavitation_amplified`]
-/// so that the single source of truth for the amplification slope lives in cfd-core.
-///
-/// Bubble collapse generates micro-jets (localised shear amplification) and
-/// pressure shockwaves that damage RBC membranes independently of macroscopic
-/// steady shear.
-///
-/// Conservative model: 3× amplification at full cavitation potential
-/// (`cav_potential = 1.0`).
-///
-/// ```text
-///   HI_amplified = base_hi × (1 + 3 × cav_potential.clamp(0, 1))
-/// ```
-///
-/// # Arguments
-///
-/// * `base_hi`          — baseline Giersiepen HI (from [`giersiepen_hi`])
-/// * `cav_potential`    — cavitation potential ∈ [0, 1]; 0 = no cavitation
-///
-/// # Example
-///
-/// ```
-/// use aequitas::systems::si::quantities::{Pressure, Time};
-/// use cfd_1d::{cavitation_amplified_hi, giersiepen_hi};
-/// let base = giersiepen_hi(Pressure::from_base(100.0), Time::from_base(1.0));
-/// let amplified = cavitation_amplified_hi(base, 1.0);
-/// assert!((amplified - base * 4.0).abs() < 1e-15);  // 1 + 3×1.0 = 4×
-/// ```
-#[inline]
-#[must_use]
-pub fn cavitation_amplified_hi(base_hi: f64, cav_potential: f64) -> f64 {
-    cfd_core::physics::hemolysis::HemolysisModel::cavitation_amplified(base_hi, cav_potential)
-}
-
-// ── Taskin (2012) strain-based hemolysis model ──────────────────────────────
-
-/// Taskin (2012) calibration constant `C_T = 1.228 × 10⁻⁵`.
-pub const TASKIN_C: f64 = 1.228e-5;
-
-/// Taskin (2012) shear stress exponent `β_T = 1.9918`.
-///
-/// Slightly different from the Giersiepen value of 1.991, reflecting the
-/// re-calibration against Lagrangian particle tracking data.
-pub const TASKIN_BETA: f64 = 1.9918;
-
-/// Taskin (2012) strain-based haemolysis index (single-segment evaluation).
-///
-/// Computes the haemolysis index using the Taskin integral-form model for a
-/// single channel segment with constant shear stress:
-///
-/// ```text
-/// HI_Taskin = C_T · τ^β_T · t
-/// ```
-///
-/// This is the single-segment (constant shear) evaluation of the full
-/// integral form:
-///
-/// ```text
-/// HI_Taskin = C_T · ∫ τ(t)^β_T dt
-/// ```
-///
-/// ## Theorem: Strain-Rate Path Dependence (Taskin et al. 2012)
-///
-/// The Taskin model captures cumulative damage along the flow path by integrating
-/// the instantaneous shear-stress contribution over time, rather than using a
-/// single power-law evaluation as in Giersiepen (1990). This integral form
-/// correctly accounts for varying shear exposure histories and yields more
-/// accurate hemolysis predictions for complex flow geometries where shear stress
-/// varies along particle trajectories.
-///
-/// For constant shear stress, the integral reduces to `C_T · τ^β_T · t`, which
-/// differs from the Giersiepen model (`C · t^α · τ^β`) in that the time
-/// dependence is linear rather than sub-linear (`α = 0.765` in Giersiepen).
-///
-/// ## Reference
-///
-/// Taskin, M. E. et al. (2012). Evaluation of Eulerian and Lagrangian Models
-/// for Hemolysis Estimation. *ASAIO J.*, 58(4), 363–372.
-///
-/// # Arguments
-///
-/// * `shear_stress` — wall shear stress \[Pa]
-/// * `exposure_time` — exposure duration \[s]
-///
-/// # Returns
-///
-/// Haemolysis index (dimensionless). Returns `0.0` for non-positive inputs.
-#[inline]
-#[must_use]
-pub fn taskin_hi(shear_stress: Pressure, exposure_time: Time) -> f64 {
-    let shear_pa = shear_stress.into_base();
-    let duration_s = exposure_time.into_base();
-    if shear_pa <= 0.0 || duration_s <= 0.0 {
-        return 0.0;
-    }
-    TASKIN_C * shear_pa.abs().max(1e-30).powf(TASKIN_BETA) * duration_s
-}
-
-// ── Convenience struct ────────────────────────────────────────────────────────
-
-/// Composite haemolysis exposure combining shear stress, duration, and
-/// cavitation potential.
-///
-/// Use [`HemolysisExposure::compute_index`] to obtain the amplified HI for
-/// a single exposure event.
-#[derive(Debug, Clone, Copy, PartialEq)]
-pub struct HemolysisExposure {
-    /// Wall shear stress at the exposure site \[Pa].
-    pub shear: Pressure,
-    /// Duration of the exposure event \[s].
-    pub duration: Time,
-    /// Local cavitation potential ∈ [0, 1]; 0 for non-venturi regions.
-    pub cavitation_potential: f64,
-}
-
-impl HemolysisExposure {
-    /// Create a new exposure event.
-    #[must_use]
-    pub fn new(shear: Pressure, duration: Time, cavitation_potential: f64) -> Self {
-        Self {
-            shear,
-            duration,
-            cavitation_potential,
-        }
-    }
-
-    /// Create an exposure event with no cavitation contribution.
-    #[must_use]
-    pub fn shear_only(shear: Pressure, duration: Time) -> Self {
-        Self::new(shear, duration, 0.0)
-    }
-
-    /// Compute the amplified haemolysis index for this exposure event.
-    ///
-    /// Combines [`giersiepen_hi`] with [`cavitation_amplified_hi`].
-    #[must_use]
-    pub fn compute_index(&self) -> f64 {
-        cavitation_amplified_hi(
-            giersiepen_hi(self.shear, self.duration),
-            self.cavitation_potential,
-        )
-    }
-}
-
-// ── Sonosensitizer Activation Kinetics (Rosenthal 2004) ──────────────────────
-
-/// Activation rate constant for Hematoporphyrin \[s⁻¹\] (Rosenthal 2004).
-pub const SENSITIZER_K_ACT_HEMATOPORPHYRIN: f64 = 0.5;
-
-/// Activation rate constant for Chlorin e6 \[s⁻¹\].
-pub const SENSITIZER_K_ACT_CHLORIN_E6: f64 = 0.8;
-
-/// Sonosensitizer activation efficiency as a function of transit time.
-///
-/// ## Theorem — First-Order Activation Kinetics (Rosenthal 2004)
-///
-/// The fraction of sonosensitizer activated during transit through a
-/// cavitation zone is governed by first-order kinetics:
-///
-/// ```text
-/// η_act = 1 − exp(−k_act · I_cav · t_transit)
-/// ```
-///
-/// where:
-/// - k_act ≈ 0.1–1.0 s⁻¹ (activation rate constant, drug-dependent)
-/// - I_cav = cavitation intensity (dimensionless, 0–1)
-/// - t_transit = residence time in cavitation zone \[s\]
-///
-/// For short transits (k·I·t << 1): η_act ≈ k·I·t (linear regime)
-/// For long transits (k·I·t >> 1): η_act → 1 (saturation)
-///
-/// **Reference**: Rosenthal, I., Sostaric, J.Z. & Riesz, P. (2004).
-/// "Sonodynamic therapy – a review of the synergistic effects of drugs
-/// and ultrasound", *Ultrason. Sonochem.* 11:349-363.
-///
-/// # Arguments
-///
-/// * `k_act` — activation rate constant \[s⁻¹\]
-/// * `cavitation_intensity` — I_cav (0–1)
-/// * `transit_time_s` — residence time in cavitation zone \[s\]
-///
-/// # Returns
-///
-/// Activation fraction η_act ∈ \[0, 1\].
-///
-/// # Example
-///
-/// ```
-/// use cfd_1d::sonosensitizer_activation_efficiency;
-/// let eta = sonosensitizer_activation_efficiency(0.5, 0.8, 2.0);
-/// assert!(eta > 0.0 && eta < 1.0);
-/// ```
-#[inline]
-#[must_use]
-pub fn sonosensitizer_activation_efficiency(
-    k_act: f64,
-    cavitation_intensity: f64,
-    transit_time_s: f64,
-) -> f64 {
-    (1.0 - (-k_act * cavitation_intensity * transit_time_s)
-        .max(-500.0)
-        .exp())
-    .clamp(0.0, 1.0)
-}
-
-// ── Rayleigh-Plesset Cavitation Bubble Models (Rayleigh 1917) ────────────────
-
-/// Empirical collapse amplification coefficient α ≈ 3.0.
-pub const RAYLEIGH_ALPHA: f64 = 3.0;
-
-/// Standard atmospheric pressure \[Pa\] used as reference for amplification.
-pub const P_REF_ATMOSPHERIC: f64 = 101_325.0;
-
-/// Rayleigh inertial collapse time for a cavitation bubble \[s\].
-///
-/// ## Theorem — Rayleigh Collapse (Rayleigh 1917)
-///
-/// For a spherical bubble collapsing from radius R_max under external
-/// pressure p_inf (neglecting gas content and surface tension):
-///
-/// ```text
-/// t_collapse = 0.915 · R_max · √(ρ / p_inf)
-/// ```
-///
-/// **Reference**: Rayleigh, Lord (1917). "On the pressure developed in a
-/// liquid during the collapse of a spherical cavity",
-/// *Phil. Mag.* 34:94-98.
-///
-/// # Arguments
-///
-/// * `r_max` — maximum bubble radius \[m\]
-/// * `rho` — liquid density \[kg/m³\]
-/// * `p_inf` — external (driving) pressure \[Pa\]
-///
-/// # Example
-///
-/// ```
-/// use cfd_1d::rayleigh_collapse_time;
-/// let t = rayleigh_collapse_time(50e-6, 1060.0, 101325.0);
-/// assert!(t > 0.0);
-/// ```
-#[inline]
-#[must_use]
-pub fn rayleigh_collapse_time(r_max: f64, rho: f64, p_inf: f64) -> f64 {
-    if r_max <= 0.0 || rho <= 0.0 || p_inf <= 0.0 {
-        return 0.0;
-    }
-
-    0.915 * r_max * (rho / p_inf).sqrt()
-}
-
-/// Micro-jet velocity from Rayleigh bubble collapse \[m/s\].
-///
-/// The collapse generates a micro-jet with velocity:
-///
-/// ```text
-/// v_jet ≈ √(2 · p_inf / ρ)
-/// ```
-///
-/// # Arguments
-///
-/// * `p_inf` — external pressure \[Pa\]
-/// * `rho` — liquid density \[kg/m³\]
-///
-/// # Example
-///
-/// ```
-/// use cfd_1d::collapse_jet_velocity;
-/// let v = collapse_jet_velocity(101325.0, 1060.0);
-/// assert!(v > 10.0, "Jet velocity at atmospheric pressure should exceed 10 m/s");
-/// ```
-#[inline]
-#[must_use]
-pub fn collapse_jet_velocity(p_inf: f64, rho: f64) -> f64 {
-    if p_inf <= 0.0 || rho <= 0.0 {
-        return 0.0;
-    }
-
-    (2.0 * p_inf / rho).sqrt()
-}
-
-/// Hemolysis amplification factor from cavitation bubble collapse.
-///
-/// The amplification scales with collapse intensity:
-///
-/// ```text
-/// A_collapse = 1 + α · (R_max / R_0)² · (p_inf / p_ref)
-/// ```
-///
-/// where α ≈ 3.0 (empirical), R_0 is equilibrium radius,
-/// p_ref = 101 325 Pa (1 atm).
-///
-/// # Arguments
-///
-/// * `r_max` — maximum bubble radius \[m\]
-/// * `r_0` — equilibrium bubble radius \[m\]
-/// * `p_inf` — external pressure \[Pa\]
-///
-/// # Example
-///
-/// ```
-/// use cfd_1d::cavitation_hemolysis_amplification;
-/// let a = cavitation_hemolysis_amplification(100e-6, 50e-6, 101325.0);
-/// assert!(a > 1.0, "Amplification should exceed unity");
-/// ```
-#[inline]
-#[must_use]
-pub fn cavitation_hemolysis_amplification(r_max: f64, r_0: f64, p_inf: f64) -> f64 {
-    if r_max <= 0.0 || r_0 <= 0.0 || p_inf <= 0.0 {
-        return 1.0;
-    }
-
-    1.0 + RAYLEIGH_ALPHA * (r_max / r_0).powi(2) * (p_inf / P_REF_ATMOSPHERIC)
-}
+pub use dynamics::{
+    cavitation_hemolysis_amplification, collapse_jet_velocity, rayleigh_collapse_time,
+    HemolysisExposure, P_REF_ATMOSPHERIC, RAYLEIGH_ALPHA,
+};
+pub use models::{
+    cavitation_amplified_hi, giersiepen_hi, sonosensitizer_activation_efficiency, taskin_hi,
+    SENSITIZER_K_ACT_CHLORIN_E6, SENSITIZER_K_ACT_HEMATOPORPHYRIN, TASKIN_BETA, TASKIN_C,
+};
 
 // ── Tests ─────────────────────────────────────────────────────────────────────
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use aequitas::systems::si::quantities::{Pressure, Time};
 
     fn shear(value: f64) -> Pressure {
         Pressure::from_base(value)
